[219] add site_prism
SitePrism is a PageObject system that uses Capybara for fluent interfaces to
pages. It's not a lot of work to add while a project is still small, so I
figured it'd be worth getting it in now.

The reason a few other gems got updated along the way is the gh gem (required by
travis) was holding some gems back to old versions (case in point, the most
recent versions of gh require activesupport ~> 5.0, blocking Rails 6!). This was
locking SitePrism to v1.3.0 (v3.4.2 is current). A happy accident ended up
down-reving gh which apparently allows for the latest versions of everything
else.