Setup temporary redirect for the home page
Currently connectivity content is being displayed on the home page.
The Central Service team would like to reclaim https://get-help-with-tech.education.gov.uk/
for the new home page which will inform users about all the get help with
tech services and support. It was aggreed and approved that connectivity
content would now be found at https://get-help-with-tech.education.gov.uk/internet-access

This temporary redirect will be removed once the new home page is ready
to be published.