Set to use major 12
We shouldn't stress about minor/patch versions of node. This commit
relaxes the checks so that we only rely on developers using 12.x

This also helps fix issues with Heroku not picking up and using the correct
minor version of node and completely failing the deploy.

Review apps should work once this has been merged.