Validate user email address in line with GOV.UK Notify rules
This change introduces user email address validation. The approach
uses a regular expression that matches what's used within GOV.UK Notify,
thereby reducing the risk that we allow email addresses that we can't
subsequently deliver emails to because they're rejected by Notify.

This code is lifted from DfE's 'Apply for teacher training' service:

https://github.com/DFE-Digital/apply-for-teacher-training

The code was introduced into Apply via this pull request:

https://github.com/DFE-Digital/apply-for-teacher-training/pull/960

which contains more context. The specs are based on the validation tests
that the Notify service used as of Feb 2018:

https://github.com/alphagov/notifications-utils/blob/e428bdd0ca82f7b97833e707eea95dae849713fb/tests/test_recipient_validation.py#L103-L149