Remove the 'Will you order' step in the school wizard
We want to avoid the risk of the school ending up without anyone
who can place orders. Since this is quite easy to do if the first
school user opts out of having a TechSource account and then doesn't
add anyone else.

It will still be possible for the first school user to opt out of having
the TechSource account by editing their account.