Harden the school details summary list component
It's possible (although highly unlikely) that an RB could access the details
of a school before they've decided whether device orders will be placed
centrally or devolved. In that case, PreorderInformation hasn't been
defined for that school yet and the component throws an exception.

It's worth fixing this particular edge-case because it's much more likely
to crop up in support, which uses a subclass of the responsible body
school details summary list.