Replace support nav links with a support homepage
The primary method of navigating to the various support functions has
been through the small links in the page header. This commit replaces
this mechanism with a support homepage which matches the designs for
schools and responsible body users.

This design allows to better explain and group functionality for the benefit
of support users who may not already be familiar with it.