Fix flaky unit test
This spec failed in the build:

https://github.com/DFE-Digital/get-help-with-tech/runs/1148408135

The spec makes assumptions about how the full name is split, but the name
is generated by faker and doesn't split nicely into first_name and last_name in all
cases.

By specifying an easy name to split, this should ensure that the spec doesn't flake.