Replacing custom breadcrumb component
`govuk-components` gem includes a breadcrumb component. We should use this
instead of maintaining our own custom one. This commit also replaces all
the hardcoded breadcrumbs and standardises the breadcrumbs.

I've still kept the breadcrumb view helper method for syntactic sugar.
Inside this method I had to transform the array that comes in as it can
include a hash (key is breadcrumb link text and value is breadcrumb path)
and a string (last breadcrumb in the list and not usually linked to).