Rework enabling_orders_for_a_school_spec
The nested spec style would have been useful while test-driving the functionality
out, but we've ended up with 8 levels of nesting which:

- are hard to follow
- contain a lot of repetition of the same steps/screens that makes execution time slow

The spec descriptions/contexts are worded in terms of UI elements, not in terms of user intent.
This makes it difficult to understand, at a glance, which use cases are covered and which aren't.

This commit:

- flattens the specs with Cucumber-like steps that encapsulate the detail
- re-cuts the flows by user intent (which, for instance, showed that important cases, like
  allowing a school to order their full allocation, or turning off ordering, weren't covered by the spec)