Introduce a spec helper for stubbing the outgoing Computacenter API
Settings.computacenter.outgoing_api.endpoint is a global setting. Specs
should ensure they don't modify this setting during setup and execution,
otherwise it affects the outcome of other specs.

As a rule, it should be on in the specs (like it is in production). Any resultant
network calls should be stubbed out via the newly introduced spec helper. If it's
necessary to change this setting, the spec should put it back to its original value.