Simplify the internet pilot service performace dashboard
The BT pilot has finished and we don't need to see the stats for that anymore.
The pilot engagement/sign-in stats are meaningless because devices and internet
is one service and currently it's not possible to tell whether a user is signing
in because of the internet pilot or because of the devices offer.

The extra mobile data requests information is still relevant so that's still available.