Consolidate the connectivity and devices service performance pages
The direction of travel for the service is to have one coherent set
of services being offered to all the users, rather than viewing devices
and connectivity as completely separate things.

The connectivity performance page had only the MNO offer stats and it
felt like overkill keeping that page around as a separate thing. When
connectivity grows, it may make sense to split these out again.

Even though the code is consolidated, tabs are used to logically split
the dashboards until the dashboards and indicators themselves are unified.