Support console: deduplicate user destroy action
Before this commit, there were two almost identical #destroy user actions, one on
Support::ResponsibleBodies::UsersController and the other on Support::Schools::UsersController.

Deletion of a user is global, it doesn't have a school or a responsible body context so I've
moved the action to Support::UsersController. This is cleaner, and also a helpful step towards
introducing an interstitial prior to deleting the user, which can now be done only in one place.