Fix zendesk ticket subject output
When a ticket is created we build up the subject based on the options
a user selects from the get support contact form.

Subject seems to be formated correctly for all users where they have
selected school,LA,multi-academty trust, college but not when
"I'm none of the above" has been chosen. The subject was being created
as "ONLINE FORM -". This commit now makes sure that the subject is set to
"ONLINE FORM - Other"