Replace reference to the email address
We are starting to roll out the new contact form. This doesn't cover all
references but it is a start. There are instances were we tag the email
subject with a particular reference and to ensure that we don't break any
zendesk macros/process, it was decided that we would't touch this for the
time being. It might be better if we focus on building forms for those
pages where we can validate the information and then create specific
zendesk tickets from those requests.