Fix issue where users use browser back buttons
Issue https://sentry.io/organizations/dfe-get-help-with-tech/issues/2138267573/?project=5320558

I managed to replicate this error by going through the full "Get support"
journey right up to the point where a zendesk ticket is created. After
I saw the "Thank you" page, I clicked on the browser back button and saw
this error.

What happens is as a user goes through the get support wizard we store the
form data in session (`session[:support_ticket]`) but once the zendesk api
has been used we clear the session. So using the back button or trying to
access other parts of the wizard may raise similiar issues.

The solution I came up with is to add a `before_action` across all the
controllers which checks if the `session[:support_ticket]` is present. if
its not then we redirect the user back to the start page.