Start validating ExtraMobileDataRequestSpreadsheet
It's only possible to extract requests out of a spreadsheet that can be
parsed and has a worksheet with a particular name. Rather than the import
mechanism trying and throwing an error, validate that these pre-conditions
are met upfront, using the ActiveModel::Validations mechanism on ExtraMobileDataRequestSpreadsheet.

BulkUploadForm also becomes invalid if the uploaded spreadsheet is invalid, meaning the
controllers that invoke it don't continue with the importing.