Prevent MNO request spreadsheet upload crash from invalid contract types
Before this commit, passing in an invalid contract type in an MNO request
spreadsheet row raised an ArgumentError.

This commit ensures that any invalid contract types are turned into nulls.
This means that the ExtraMobileDataRequest is still invalid because contract
type can't be null, but at least now the spreadsheet importer can skip that
row and import the others.