Fix missing translation issue when a MNO request spreadsheet cannot be processed
There was an i18n string that was being used in two similar controllers. One controller
was updated and the string moved, but I forgot to update the other controller, so users
are seeing a 'missing translation' error.

I've added spec coverage around both controllers to prevent this from regressing in future.