Allow MNO requests with invalid phone numbers to be updated
PR https://github.com/DFE-Digital/get-help-with-tech/pull/1118 tightened
up validation of device phone numbers so that some phone numbers (numbers
starting with 07xxx but which weren't mobile numbers) suddenly became invalid.

Because the validation triggered for updates as well, an MNO user wasn't able to
update a request with an invalid phone number:

https://sentry.io/organizations/dfe-get-help-with-tech/issues/2154600108/?environment=production&project=5320558

This commit changes the validation to only fire on create.