Ensure only one validation error when an MNO request has no valid network
Currently, if mobile_network isn't set, then the record ends up with two
validation errors: one on the mobile_network association, and another on
mobile_network_id (because we have an explicit validation on that field).

We need to preserve the mobile_network_id validation because it's a workaround
for correctly rendering the form error on the manual MNO request submission form.
So this commit makes the mobile_network association optional. It also adds spec
coverage to ensure that it's not possible to create an MNO request without the network.