Remove requests for specific circumstances
Requests for specific circumstances were to allow schools to order devices for children not at school before disruption is reported – for example clinically extremely vulnerable children.

Now ordering is open for all schools, the "get devices early for a subset" feature doesn't apply.

Allocations do not cover years 1 and 2, and children in those years are still eligible for devices. We will use public guidance to describe this process – many schools will have 0 allocations so might not be on service.

- Remove all links and templates for making a request for specific circumstances
- Leave in place "can order for specific circumstances" status for now