Surface the ticket number in the zendesk response
When we call ZendeskAPI::Request.create method, zendesk response to
us include all the information we supplied as well as the ticket reference
number (i.e the id). This is really useful for our users to make them
feel more confident that their request has been received.

Later we will display this ticket number to the user in another commit