Reuse reigster users details and store in session
When users contact us using the contact form, we ask them for their
full name, email and telephone number. All of this details we have for
registered users. So to help them to complete the form, we can just use
the information which we hold for them if they have signed into the service.

We also build & store the url that the support agents can use to access the
persons profile from the support portal.