Pass additional information to zendesk_service
As we are automatically creating zendesk tickets using the api and we
have a path to the users profile. It would be good for support agents
to be able to quickly access the users profile by visiting this link from
zendesk. This will save them having to try go through various screens to
find this information.