Setup one import script to setup local db
Currently, there are a large number of import tasks used to
import various sets of data. These also have to run in a specific
order as some of the imports are dependent on other sets of data.

Developers shoudl now be able to run `bundle exec rake import:all`
and the task should be able to import and setup all the data
needed for local developement database.