Disable worker in Heroku review apps
When a Heroku review app lives for more than a few hours, it can
trigger the overnight GIAS updates. This brings in up to 70k rows
to the DB, and triggers warnings from Heroku about the database
size exceeding the maximum available on the free plan.

Since we are now flashing login token links to the user on Heroku,
there's no reason that we can think of for actually keeping the
worker process for review apps.

This PR removes the worker process.