Refactor guidance_page_path to root_path
`guidance_page_path` is very confusing for dev who might be new to the
service as it is really just the service home page and not a link to
specific guidance. This refactor makes it more obvious that this is the
starting point/home page of the service.

There is another piece of work goign on which will rename the breadcrumbs
copy to "Home".