Reorganise "Guides for parents, carers" nav
We have decided to move the list Guides that parents,carers, pupils and
students might find usefull from the bottom of the devices guidance page
to a new second level page which can be navigated to from the home page.

THe guides themselves have not been updated onlythe way in which users
may locate the guides have.