Refactor devices guidance features
This change should hopefully make it easier to understand in the future.
"Audience" and the key "responsible_body_user" made this feature sound like
it was used to restrict the content to a particular user and that they
would have needed to be signed in to access it. In fact, it was being used
to group the guidances into sections and then play them back in
a particular order.