Remove FactoryBot from email audit preview
Generally speaking component previews should be static examples and shouldn't
care how the data is structured. In this commit, I've replaced FactoryBot
with mocked out data set. Previews that use FactoryBot will not work
if the app is running in Producation mode which is what our review apps
and GOV.UK Paas dev instance is doing. This is because FactoryBot gem
is only installed for development & test.