Tactical fix to allow allocations to be decreased
… for schools that are not in a virtual cap pool.

Note: I've gone for a "light touch" fix here that doesn't attempt to change too much – there's a longer term piece required for refactoring + code consolidation + potential fixes, later on, after we figure out / confirm what the intended product design for allocations and vcap pools is moving forward.

Changes here:

- Allow allocations to be decreased, but only for schools that are not in a virtual cap pool.
- For schools that are in a virtual cap pool, we a) show a warning message in the UI that decreasing is not allowed, and b) have a validation (at the form model layer) to prevent decreasing.
- The allocation now has a minimum, set to the current `SchoolDeviceAllocation#raw_devices_ordered`, via a validation in the form model. The UI has also been updated to show a hint.
  - Note: this validation is only on the form model layer as we only want to enforce it for the support user flow. See <https://trello.com/c/FxejJUlY/716-remove-the-validation-that-devicesordered-cannot-be-greater-than-cap> for context on why this is not a model validation.
- Updated existing specs + new specs for the `AllocationForm` model.

Note an edge case here: because the main change here mirrors the cap adjustment logic in `SchoolAllocationDevice#cap_implied_by_order_state`, the cap is not adjusted for schools that have an order state of `can_order_for_specific_circumstances`, which could result in the same HTTP 422 error bug hitting if an allocation is decreased below the cap for a school with this order state.